Move tests to repository root
