+++ tests/templates/wagtailmedia/media/add.html
@@ -1,0 +1,3 @@
+{% extends "wagtailmedia/media/add.html" %}
+
+{% block action %}/somewhere/else{% endblock %}
