+++ tests/testapp/forms.py
@@ -1,0 +1,15 @@
+from django.forms import ModelForm
+from django.forms.widgets import Widget
+
+
+class OverridenWidget(Widget):
+    pass
+
+
+class AlternateMediaForm(ModelForm):
+    class Meta:
+        widgets = {
+            "tags": OverridenWidget,
+            "file": OverridenWidget,
+            "thumbnail": OverridenWidget,
+        }
