+++ tests/testapp/templates/wagtailmedia_tests/event_page.html
