+++ tests/testextends/templates/wagtailmedia/media/index.html
@@ -1,0 +1,3 @@
+{% extends "wagtailmedia/media/index.html" %}
+
+{% block add_actions %}You shan't act{% endblock %}
