--- wagtailmedia/tests/templates/wagtailmedia/media/add.html
@@ -1,3 +1,0 @@
-{% extends "wagtailmedia/media/add.html" %}
-
-{% block action %}/somewhere/else{% endblock %}
